Added basic demo to compile + row assets
To compile, download 4D Workshop, choose VisiGenie. You need to compile
to the µSD card, then plug it in via USB to test. It is best to do this
in two different steps, because the adapter gets really warm when
plugged in for more than a few minutes and this lets it cool down.

Actual additional functionality not implemented yet- just a demo to
display the grid on the touchscreen